Created class RecordManager
Created singleton class RecordManager.
This class provides methods to facilitate record management
(record creation, update, deletion, etc.) with internal
invenio record API.
