+++ invenio_rdm_pure/source/rdm/add_record.py
@@ -12,20 +12,6 @@ import os.path
 import time
 from datetime import date
 
-from flask import current_app
-from flask_principal import Identity
-from invenio_access.permissions import any_user, superuser_access
-from invenio_db import InvenioDB, db
-from invenio_rdm_records.records import BibliographicRecord
-from invenio_rdm_records.services import (
-    BibliographicRecordService,
-    BibliographicRecordServiceConfig,
-)
-from invenio_records import InvenioRecords, Record
-from invenio_records_permissions.generators import AnyUser
-from invenio_records_permissions.policies import RecordPermissionPolicy
-from invenio_records_resources.services.records.results import RecordItem
-
 from ...setup import (
     accessright_pure_to_rdm,
     data_files_name,
@@ -54,25 +40,6 @@ from ..rdm.requests_rdm import Requests
 from ..rdm.run.groups import RdmGroups
 from ..rdm.versioning import Versioning
 from ..reports import Reports
-
-
-class PermissionPolicy(RecordPermissionPolicy):
-    """Policy class to use with BibliographicRecordServiceConfig."""
-
-    can_create = [AnyUser()]
-    can_update_files = [AnyUser()]
-    can_publish = [AnyUser()]
-    can_read = [AnyUser()]
-    can_update = [AnyUser()]
-    can_delete = [AnyUser()]
-    # FIXME: can_delete has to be changed as soon as user group management
-    # works with "current_app.extensions["security"].datastore"
-
-
-class ServiceConfig(BibliographicRecordServiceConfig):
-    """Config class to use with BibliographicRecordService."""
-
-    permission_policy_cls = PermissionPolicy
 
 
 class RdmAddRecord:
@@ -200,60 +167,6 @@ class RdmAddRecord:
 
         # Updates the versioning data of all records with the same uuid
         self._update_all_uuid_versions()
-
-    def create_record(self, data) -> RecordItem:
-        """Creates a record from JSON data."""
-        identity = Identity(current_app.config.get(self.rdm_db.get_pure_user_id()))
-        identity.provides.add(any_user)
-        service = BibliographicRecordService(config=ServiceConfig)
-        record = service.create(identity, data)
-        service.publish(id_=record.id, identity=identity)
-
-        if record is not None:
-            return record
-        else:
-            return None
-
-    def update_record(self, recid: str, data) -> RecordItem:
-        """Updates a record with JSON data."""
-        identity = Identity(current_app.config.get(self.rdm_db.get_pure_user_id()))
-        identity.provides.add(any_user)
-        service = BibliographicRecordService(config=ServiceConfig)
-        original_record = service.read(id_=recid, identity=identity)
-        original_revision_id = original_record._record.revision_id
-        updated_record = service.update(id_=recid, identity=identity, data=data)
-        # service.publish(id_=record.id, identity=identity)
-
-        if (
-            updated_record is None
-            or updated_record._record.revision_id != original_revision_id + 1
-        ):
-            raise RuntimeError("Failed to update record.")
-        else:
-            return updated_record
-
-    def delete_record(self, recid: str) -> None:
-        """Deletes record with given recid."""
-        if not recid:
-            raise ValueError("Can't delete record without providing recid.")
-        identity = Identity(current_app.config.get(self.rdm_db.get_pure_user_id()))
-        identity.provides.add(any_user)
-        service = BibliographicRecordService(config=ServiceConfig)
-        deleted = service.delete(id_=recid, identity=identity)
-        if not deleted:
-            raise RuntimeError("Failed to delete record.")
-
-    def is_newest_record(self, record: RecordItem) -> bool:
-        """Checks if the given record is the most recently inserted one."""
-        if record is not None:
-            newest_query = {"sort": "newest", "size": 1, "page": 1}
-            response = self.rdm_requests.get_metadata(newest_query)
-            if response.ok:
-                response_json = json.loads(response.text)
-                newest_recid = response_json["hits"]["hits"][0]["id"]
-                if record.id == newest_recid:
-                    return True
-        return False
 
     def _access_right_and_restrictions(self, item):
         """Description."""
